passwordreset package reorganized
