+++ backend/src/main/scala/com/softwaremill/bootzooka/ServerConfig.scala
@@ -1,0 +1,18 @@
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
+
+package com.softwaremill.bootzooka
+
+import com.typesafe.config.Config
+
+trait ServerConfig {
+  def rootConfig: Config
+
+  lazy val serverHost: String = rootConfig.getString("server.host")
+  lazy val serverPort: Int = rootConfig.getInt("server.port")
+}
