--- backend/src/main/scala/com/softwaremill/bootzooka/config/ServerConfig.scala
@@ -1,10 +1,0 @@
-package com.softwaremill.bootzooka.config
-
-import com.typesafe.config.Config
-
-trait ServerConfig {
-  def rootConfig: Config
-
-  lazy val serverHost: String = rootConfig.getString("server.host")
-  lazy val serverPort: Int = rootConfig.getInt("server.port")
-}
