+++ backend/src/main/scala/com/softwaremill/bootzooka/passwordreset/api/PasswordResetRoutes.scala
@@ -1,8 +1,17 @@
-package com.softwaremill.bootzooka.passwordreset
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
+
+package com.softwaremill.bootzooka.passwordreset.api
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Directives._
 import com.softwaremill.bootzooka.common.api.RoutesSupport
+import com.softwaremill.bootzooka.passwordreset.application.PasswordResetService
 import com.softwaremill.bootzooka.user.api.SessionSupport
 import io.circe.generic.auto._
