+++ backend/src/main/scala/com/softwaremill/bootzooka/passwordreset/application/PasswordResetService.scala
@@ -1,10 +1,19 @@
-package com.softwaremill.bootzooka.passwordreset
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
+
+package com.softwaremill.bootzooka.passwordreset.application
 
 import java.time.Instant
 
+import com.softwaremill.bootzooka.CoreConfig
 import com.softwaremill.bootzooka.common.Utils
-import com.softwaremill.bootzooka.config.CoreConfig
 import com.softwaremill.bootzooka.email.{EmailContentWithSubject, EmailService, EmailTemplatingEngine}
+import com.softwaremill.bootzooka.passwordreset.domain.PasswordResetCode
 import com.softwaremill.bootzooka.user.application.UserDao
 import com.softwaremill.bootzooka.user.domain.User
 import com.typesafe.scalalogging.StrictLogging
