+++ backend/src/main/scala/com/softwaremill/bootzooka/passwordreset/domain/PasswordResetCode.scala
@@ -1,7 +1,15 @@
-package com.softwaremill.bootzooka.passwordreset
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
 
-import java.time.{Instant, OffsetDateTime, ZoneOffset}
+package com.softwaremill.bootzooka.passwordreset.domain
+
 import java.time.temporal.ChronoUnit
+import java.time.{Instant, OffsetDateTime, ZoneOffset}
 import java.util.UUID
 
 import com.softwaremill.bootzooka.user.domain.User
